Attack animation WIP
Going to rip up the current animation loop, wish me luck
