+++ Lasso/Assets/Animations/attack.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a00545a7c7cb39143af170b47438e3fb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
