--- Lasso/Assets/PlayerAttack.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PlayerAttack : MonoBehaviour
-{
-    [SerializeField] private float attackCooldown;
-    private Animator anim;
-    private PlayerMovement playerMovement;
-
-    private void Awake()
-    {
-
-	}
-}
